Updates instructions to remove Fritz!Box error
Previous instructions resulted in the Fritz!Box generating error messages
because it was unable to validate that $domain was updated. It was updating
$subdomain.$domain instead, but attempting to validate with $domain. Now the
instructions correctly specify how to configure the Fritz!Box, by setting the
"domain" string to $subdomain.$domain and embedding both of the strings into
the Url parameters when invoking this service.

Fixes #2.